+++ app/services/statistics/builders/reservations_builder_service.rb
@@ -21,16 +21,34 @@ class Statistics::Builders::ReservationsBuilderService
             stat[:stat] = (type == 'booking' ? 1 : r[:nb_hours])
             stat["#{category}Id".to_sym] = r["#{category}_id".to_sym]
 
-            if category == 'event'
-              stat[:eventDate] = r[:event_date]
-              stat[:eventTheme] = r[:event_theme]
-              stat[:ageRange] = r[:age_range]
-            end
-
+            stat = add_custom_attributes(category, stat, r)
             stat.save
           end
         end
       end
+    end
+
+    def add_custom_attributes(category, stat, reservation_data)
+      stat = add_event_attributes(category, stat, reservation_data)
+      add_training_attributes(category, stat, reservation_data)
+    end
+
+    def add_event_attributes(category, stat, reservation_data)
+      return stat unless category == 'event'
+
+      stat[:eventDate] = reservation_data[:event_date]
+      stat[:eventTheme] = reservation_data[:event_theme]
+      stat[:ageRange] = reservation_data[:age_range]
+
+      stat
+    end
+
+    def add_training_attributes(category, stat, reservation_data)
+      return stat unless category == 'training'
+
+      stat[:trainingDate] = reservation_data[:training_date]
+
+      stat
     end
   end
 end