+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog Fab-manager
 
+- Fix a bug: unable to export statistics
 - Fix a bug: soft destroyed machines and spaces are still reported in the OpenAPI
 
 ## v5.5.5 2022 November 22